+++ src/document.rs
@@ -5,7 +5,9 @@ use html5ever::parse_document;
 use html5ever::tree_builder;
 use html5ever::tree_builder::{ElementFlags, NodeOrText, QuirksMode, TreeSink};
 use html5ever::ParseOpts;
-use html5ever::{local_name, namespace_url, ns};
+#[allow(unused_imports)]
+use html5ever::namespace_url;
+use html5ever::{local_name, ns};
 use html5ever::{Attribute, QualName};
 
 use tendril::{StrTendril, TendrilSink};